prof_second/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
et-xmlfile==1.1.0
openpyxl==3.1.5
sqlparse==0.5.1
typing_extensions==4.12.2
